test(rebuild): isolate prepared refresh coverage
